+++ packages/components/src/Input/AddressInput.tsx
@@ -1,7 +1,9 @@
+"use client";
+
 import { blo } from "blo";
 import { Address } from "viem";
 import { useAddressInput } from "@scaffold-ui/hooks";
-import { InputBase } from "./InputBase";
+import { BaseInput } from "./BaseInput";
 import { CommonInputProps } from "./utils";
 import { useEffect, useState } from "react";
 
@@ -24,6 +26,7 @@ export type AddressInputProps = CommonInputProps<Address | string>;
  * @param {string} [props.placeholder] - (Optional) Placeholder text for the input field.
  * @param {(value: Address) => void} [props.onChange] - Callback function called when the input value changes.
  * @param {boolean} [props.disabled] - (Optional) Whether the input is disabled.
+ * @param {CSSProperties} [props.style] - (Optional) Styles for the input.
  *
  * @example
  * <AddressInput
@@ -38,7 +41,7 @@ export type AddressInputProps = CommonInputProps<Address | string>;
  *   disabled={false}
  * />
  */
-export const AddressInput = ({ value, name, placeholder, onChange, disabled }: AddressInputProps) => {
+export const AddressInput = ({ value, name, placeholder, onChange, disabled, style }: AddressInputProps) => {
   const {
     ensAddress,
     ensName,
@@ -48,8 +51,6 @@ export const AddressInput = ({ value, name, placeholder, onChange, disabled }: A
     isEnsAvatarLoading,
     isEnsAddressError,
     isEnsNameError,
-    isEnsNameSuccess,
-    isEnsAddressSuccess,
     settledValue,
   } = useAddressInput({
     value,
@@ -57,13 +58,7 @@ export const AddressInput = ({ value, name, placeholder, onChange, disabled }: A
 
   const [enteredEnsName, setEnteredEnsName] = useState<string>();
 
-  const reFocus =
-    isEnsAddressError ||
-    isEnsNameError ||
-    isEnsNameSuccess ||
-    isEnsAddressSuccess ||
-    ensName === null ||
-    ensAddress === null;
+  const reFocus = isEnsAddressError || isEnsNameError || ensName === null || ensAddress === null;
 
   // ens => address
   useEffect(() => {
@@ -79,32 +74,33 @@ export const AddressInput = ({ value, name, placeholder, onChange, disabled }: A
   }, [value]);
 
   return (
-    <InputBase<Address>
+    <BaseInput<Address>
       name={name}
       placeholder={placeholder}
       error={ensAddress === null}
       value={value as Address}
       onChange={onChange}
+      style={style}
       disabled={isEnsAddressLoading || isEnsNameLoading || disabled}
       reFocus={reFocus}
       prefix={
         ensName ? (
-          <div className="flex bg-[#dae8ff] rounded-l-full items-center">
+          <div className="flex rounded-l-full items-center bg-sui-input-border" style={style}>
             {isEnsAvatarLoading && (
-              <div className="animate-pulse bg-[#f4f8ff] w-[35px] h-[35px] rounded-full shrink-0"></div>
+              <div className="animate-pulse w-[35px] h-[35px] rounded-full shrink-0 bg-sui-input-background" />
             )}
             {ensAvatar ? (
               <span className="w-[35px]">
                 <img className="w-full rounded-full" src={ensAvatar} alt={`${ensAddress} avatar`} />
               </span>
             ) : null}
-            <span className="text-[#93bbfb] px-2">{enteredEnsName ?? ensName}</span>
+            <span className="px-2 text-sui-accent">{enteredEnsName ?? ensName}</span>
           </div>
         ) : (
           (isEnsNameLoading || isEnsAddressLoading) && (
-            <div className="flex bg-[#dae8ff] rounded-l-full items-center gap-2 pr-2">
-              <div className="animate-pulse bg-[#f4f8ff] w-[35px] h-[35px] rounded-full shrink-0"></div>
-              <div className="animate-pulse bg-[#f4f8ff] h-3 w-20"></div>
+            <div className="flex rounded-l-full items-center gap-2 pr-2 bg-sui-input-border" style={style}>
+              <div className="animate-pulse w-[35px] h-[35px] rounded-full shrink-0 bg-sui-input-background" />
+              <div className="animate-pulse h-3 w-20 bg-sui-input-background" />
             </div>
           )
         )