+++ packages/hooks/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @scaffold-ui/hooks
 
+## 0.1.2
+
+### Patch Changes
+
+- b884a95: make address optional for useBalance and Balance component
+- b884a95: fix: useAddress to only use mainnet for resolving ens
+
+## 0.1.2-rc.1
+
+### Patch Changes
+
+- 5e5778b: make address optional for useBalance and Balance component
+
+## 0.1.2-rc.0
+
+### Patch Changes
+
+- 6674b6e: fix: useAddress to only use mainnet for resolving ens
+
 ## 0.1.1
 
 ### Patch Changes